custom field fix for qaoa params, missing authenticate, renamings
