+++ src/quark_plugin_luna/luna_qa.py
@@ -12,7 +12,7 @@ from luna_quantum import LunaSolve
 from luna_quantum.translator import BqmTranslator
 from luna_quantum.algorithms import QuantumAnnealing
 
-from .utils import converter_solution, converter_model, get_runtime
+from .utils import converter_solution, converter_model, get_runtime, get_luna_api_key
 
 @dataclass
 class LUNAQA(Core):
